+++ system-tests/multiple-agents-test/src/main/java/systemundertest/SystemUnderTest.java
@@ -1,0 +1,8 @@
+package systemundertest;
+
+public class SystemUnderTest {
+
+	public int foo() {
+		return 2;
+	}
+}