+++ CHANGELOG.md
@@ -4,6 +4,31 @@ We use [semantic versioning](http://semver.org/):
 - MINOR version when you add functionality in a backwards-compatible manner, and
 - PATCH version when you make backwards compatible bug fixes.
 
+# next release
+
+# 24.1.1
+
+- [fix] Maven plugin for TIA: sometimes the agent was not attached to a Spring Boot application during integration tests
+
+# 24.1.0
+
+- [feature] The Maven plugin now writes temporary reports to `target/tia/reports` in case of upload errors so they can be inspected manually.
+- [feature] The agent logs a warning when multiple java agents are used and recommends registering the Teamscale JaCoCo Agent first.
+
+# 24.0.1
+
+- [fix] fix POM metadata of Maven plugin to allow publishing it
+
+# 24.0.0
+
+- [feature] add Maven plugin for TIA
+- [feature] added official support for Java 17 and 18 and experimental support for Java 19.
+- [feature] New command line option: `debug`. Simplifies debugging by avoiding the process of providing an XML logging
+  configuration file. In debug mode logs are written to console and a configurable directory. For more details, see
+  [here](agent/README.md#general-options).
+- [breaking change] Artifactory uploader: use of new default upload schema for easy integration with Teamscale. To keep the old behavior add the `artifactory-legacy-path=true` option. For more details see 
+  [here](agent/README.md#options-for-the-artifactory-upload).
+
 # 23.1.1
 
 - [fix] _teamscale-gradle-plugin_: The `TestImpacted` task did not execute any tests